Refactor errors into it's own errors.go file
