+++ geo/geopoint.go
@@ -1,8 +1,6 @@
 package geo
 
 import (
-	"errors"
-	"fmt"
 	"math"
 	"sync"
 )
@@ -26,18 +24,6 @@ type Config struct {
 var DefaultConfig = &Config{
 	defaultConcurrencyLimit: defaultConcurrency,
 	defaultMaxSegments:      defaultMaximumSegments,
-}
-
-// List of errors
-var (
-	ErrRadiusLargerThanEarth    = errors.New("radius can not be greater than the radius of the Earth, 6378137 metres")
-	ErrTooFewSegments           = errors.New("too few segments, this should be set to 3 or more")
-	ErrInvalidLongitudinalPoint = errors.New("longitude has to be between -180 and 180")
-	ErrInvalidLatitudinalPoint  = errors.New("latitude has to be between -90 and 90")
-)
-
-func ErrTooManySegments(maximumSegments int) error {
-	return fmt.Errorf("too many segments, this should be less than %d", maximumSegments)
 }
 
 // GeoStructure describes the shape of the geographical location
